Moved instantid pipeline to separate folder
