+++ optimum/habana/diffusers/__init__.py
@@ -15,5 +15,5 @@ from .pipelines.stable_diffusion_xl.pipeline_stable_diffusion_xl import GaudiSta
 from .pipelines.stable_diffusion_xl.pipeline_stable_diffusion_xl_img2img import GaudiStableDiffusionXLImg2ImgPipeline
 from .pipelines.stable_diffusion_xl.pipeline_stable_diffusion_xl_inpaint import GaudiStableDiffusionXLInpaintPipeline
 from .pipelines.stable_video_diffusion.pipeline_stable_video_diffusion import GaudiStableVideoDiffusionPipeline
-from .pipelines.stable_diffusion_xl.pipeline_stable_diffusion_xl_instantid import GaudiStableDiffusionXLInstantIDPipeline
+from .pipelines.instantid.pipeline_stable_diffusion_xl_instantid import GaudiStableDiffusionXLInstantIDPipeline
 from .schedulers import GaudiDDIMScheduler, GaudiEulerAncestralDiscreteScheduler, GaudiEulerDiscreteScheduler
