+++ optimum/habana/diffusers/pipelines/instantid/ip_adapter/attention_processor.py
