+++ optimum/habana/diffusers/pipelines/instantid/ip_adapter/resampler.py
