+++ optimum/habana/diffusers/pipelines/instantid/ip_adapter/utils.py
